State passed from server to client
WHY DID I EVER DOUBT MYSELF IT WORKS OF COURSE IT WORKS YAAYYY
But don't worry, the formatting is still awful so it's only a partial victory.

Just to clarify: Memory (upcoming instructions), Registers and Flags data is now queried from the server. Register data is also displayed, but the formatting is terrible because it's just a big block.