Get CI/CD version into the Java build
Hooray! I actually did it!
Gradle (and GitHub Actions by extension) will now put the VM Version string passed to Gradle into a properties file which the running VM can read! This means that there is now a single source of truth for the version of the VM: the GitHub actions configuration.