Split up 'status' API
WARNING, THERE ARE BUILD ERRORS!
I'd like to add reading debug symbols from a file, or over the network, but the '/status/' endpoint is handling too much, and the response packet is getting too large. I'm in the process of breaking it up into a '/status/' REST controller and endpoints for memory slices, registers, etc. ... and (the new guy!) debug symbols.

Also, the updating of ShellDisplayState is now done *within* ShellDisplayState. That means that (sadly) SDS is not just a record, but it does make a lot more sense to keep it together and away from ShellDisplay as the update and drawing logic become more complicated.